backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
yfinance==0.2.36
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
pymongo==4.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
